requirements.txt
astropy~=5.1
numpy~=1.23.3
joblib~=1.2.0
optuna~=3.0.3
pandas~=1.5.0
torch~=1.12.1
plotly~=5.11.0
matplotlib~=3.6.0
scikit-learn~=1.1.3
torchvision~=0.13.1
mpi4py~=3.1.4
PyYAML~=6.0
